Move Authentication interfaces
Moved Interfaces related to authentication etc to the
Infrastructure\Auth\Contracts namespace.
