+++ classes/Http/Controller/TalkController.php
@@ -5,9 +5,9 @@ namespace OpenCFP\Http\Controller;
 use OpenCFP\Application\NotAuthorizedException;
 use OpenCFP\Application\Speakers;
 use OpenCFP\Domain\Model\Talk;
-use OpenCFP\Domain\Services\Authentication;
 use OpenCFP\Http\Form\TalkForm;
 use OpenCFP\Http\View\TalkHelper;
+use OpenCFP\Infrastructure\Auth\Contracts\Authentication;
 use Swift_Message;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Security\Csrf\CsrfToken;
