+++ classes/Provider/TalkRatingProvider.php
@@ -2,9 +2,9 @@
 
 namespace OpenCFP\Provider;
 
-use OpenCFP\Domain\Services\Authentication;
 use OpenCFP\Domain\Services\TalkRating\TalkRatingContext;
 use OpenCFP\Domain\Services\TalkRating\TalkRatingStrategy;
+use OpenCFP\Infrastructure\Auth\Contracts\Authentication;
 use Pimple\Container;
 use Pimple\ServiceProviderInterface;
