+++ tests/Unit/Domain/Services/TalkRating/TalkRatingTest.php
@@ -4,9 +4,9 @@ namespace OpenCFP\Test\Unit\Domain\Services\TalkRating;
 
 use Mockery;
 use OpenCFP\Domain\Model\TalkMeta;
-use OpenCFP\Domain\Services\Authentication;
 use OpenCFP\Domain\Services\TalkRating\TalkRatingException;
 use OpenCFP\Domain\Services\TalkRating\YesNoRating;
+use OpenCFP\Infrastructure\Auth\Contracts\Authentication;
 
 /**
  * We Use the YesNoRating class to test the base class, since we know exactly what values are allowed
