+++ tests/Unit/Infrastructure/Auth/SentinelIdentityProviderTest.php
@@ -5,8 +5,8 @@ namespace OpenCFP\Test\Unit\Infrastructure\Auth;
 use Cartalyst\Sentinel\Sentinel;
 use Mockery as m;
 use OpenCFP\Domain\Model\User;
-use OpenCFP\Domain\Services\IdentityProvider;
 use OpenCFP\Domain\Speaker\SpeakerRepository;
+use OpenCFP\Infrastructure\Auth\Contracts\IdentityProvider;
 use OpenCFP\Infrastructure\Auth\SentinelIdentityProvider;
 use OpenCFP\Test\BaseTestCase;
 use OpenCFP\Test\Helper\Faker\GeneratorTrait;
