+++ tests/Unit/Infrastructure/Auth/SpeakerAccessTest.php
@@ -3,7 +3,7 @@
 namespace OpenCFP\Test\Unit\Infrastructure\Auth;
 
 use Mockery;
-use OpenCFP\Domain\Services\Authentication;
+use OpenCFP\Infrastructure\Auth\Contracts\Authentication;
 use OpenCFP\Infrastructure\Auth\SpeakerAccess;
 use OpenCFP\Test\WebTestCase;
 use Symfony\Component\HttpFoundation\RedirectResponse;
